@@ -8,11 +8,14 @@
     "lint": "next lint",
     "start": "next start"
   },
+  "type": "module",
   "dependencies": {
+    "aws4": "^1.13.2",
+    "axios": "^1.7.9",
     "cloudinary": "^2.5.1",
     "mapbox-gl": "^3.5.1",
-    "next": "~14.2.4",
-    "next-auth": "^4.24.7",
+    "next": "^15.1.0",
+    "next-auth": "^4.24.11",
     "photoswipe": "^5.4.4",
     "react": "^18",
     "react-dom": "^18",
