@@ -1,20 +1,13 @@
-const mongoose = require('mongoose');
+import mongoose from "mongoose";
 
 const connectDB = async () => {
-  if (mongoose.connection.readyState) {
-    console.log('MongoDB is already connected...');
-    return;
-  }
-
-  try {
-    await mongoose.connect(process.env.MONGODB_URI, {
-      serverSelectionTimeoutMS: 20000, // Set server selection timeout to 20 seconds
-    });
-    console.log('MongoDB connected...');
-  } catch (error) {
-    console.error('Error connecting to MongoDB:', error.message);
-    throw new Error('Failed to connect to MongoDB.');
-  }
+	try {
+		const conn = await mongoose.connect(process.env.MONGO_URI);
+		console.log(`Connected to MongoDB: ${conn.connection.host}`);
+	} catch (error) {
+		console.error(`Error: ${error.message}`);
+		process.exit(1);
+	}
 };
 
-module.exports = connectDB;
+export default connectDB;
