@@ -1,7 +1,7 @@
 'use client';
 
-import bookmarkProperty from '@/backend/actions/bookmarkProperty';
-import checkBookmarkStatus from '@/backend/actions/checkBookmarkStatus';
+import bookmarkProperty from '@/app/actions/bookmarkProperty';
+import checkBookmarkStatus from '@/app/actions/checkBookmarkStatus';
 import { useSession } from 'next-auth/react';
 import { useEffect, useState } from 'react';
 import { FaBookmark } from 'react-icons/fa';
