@@ -1,5 +1,3 @@
-import connectDB from "@/backend/config/database";
-import User from "@/models/User";
 import GoogleProvider from "next-auth/providers/google";
 
 export const authOptions = {
@@ -17,33 +15,12 @@ export const authOptions = {
 		}),
 	],
 	callbacks: {
-		// Invoked on successful signin
 		async signIn({ profile }) {
-			// 1. Connect to database
-			await connectDB();
-			// 2. Check if user exists
-			const userExists = await User.findOne({ email: profile.email });
-			// 3. If not, then add user to database
-			if (!userExists) {
-				// Truncate user name if too long
-				const username = profile.name.slice(0, 20);
-
-				await User.create({
-					email: profile.email,
-					username,
-					image: profile.picture,
-				});
-			}
-			// 4. Return true to allow sign in
+			// Handle user database logic here
 			return true;
 		},
-		// Modifies the session object
 		async session({ session }) {
-			// 1. Get user from database
-			const user = await User.findOne({ email: session.user.email });
-			// 2. Assign the user id to the session
-			session.user.id = user._id.toString();
-			// 3. return session
+			// Modify session object here
 			return session;
 		},
 	},
