@@ -1,15 +1,15 @@
-import { authOptions } from '@/utils/authOptions';
-import { getServerSession } from 'next-auth/next';
+import { authOptions } from "./authOptions.js";
+import { getServerSession } from "next-auth/next";
 
-export const getSessionUser = async () => {
-  const session = await getServerSession(authOptions);
+export const getSessionUser = async (req, res) => {
+	const session = await getServerSession(req, res, authOptions);
 
-  if (!session || !session.user) {
-    return null;
-  }
+	if (!session || !session.user) {
+		return null;
+	}
 
-  return {
-    user: session.user,
-    userId: session.user.id,
-  };
+	return {
+		user: session.user,
+		userId: session.user.id,
+	};
 };
